R-scripts to create plots from MS source data
In the manuscript we included excel sheets with source data to create Figures 2, 3, 4; Extended Data 5, 6, 7; and the PERMANOVA analysis of pi values. Each sheet in the excel files need to be extracted into a separate file and the location PATHs entered into the R-scripts to produce the raw unmodified figures.